+++ frontend/package.json
@@ -14,6 +14,7 @@
     "axios": "^1.13.6",
     "react": "^19.2.4",
     "react-dom": "^19.2.4",
+    "react-router-dom": "^7.13.2",
     "tailwindcss": "^4.2.1"
   },
   "devDependencies": {
