+++ frontend/src/context/useAuthContext.tsx
@@ -1,0 +1,6 @@
+import { useContext } from "react"
+import AuthContext from "./AuthContext"
+
+export default function useAuthContext() {
+  return useContext(AuthContext)
+}
