+++ frontend/src/types.ts
@@ -1,0 +1,11 @@
+export interface GetDrawingsResponse {
+  message: string,
+  data: Drawing[]
+}
+
+export interface Drawing {
+  id: string,
+  name: string,
+  description: string,
+  path: string
+}
